+++ tutorial/src/CHANGELOG.md
@@ -9,3 +9,5 @@ Rename several structures to better reveal their natures.
 - remove `GrowingStrategy`
 - rename `tuning_cluster_size_limit` to `cluster_node_limit`
 - rename `PrimalDualType` to `SolverType`
+- rename `MaxUpdateLength` to `Obstacle`
+- rename `GroupMaxUpdateLength` to `DualReport`